Add asset support
