+++ uploader/src/main/scala/com/gu/media/LambdaActionHandler.scala
@@ -12,47 +12,9 @@ class LambdaActionHandler(store: UploadsDataStore, plutoStore: PlutoDataStore, a
                           youTube: YouTubeUploader, hmac: HmacRequestSupport)
   extends UploadActionHandler(store, plutoStore, aws, youTube) {
 
-  private val domain = aws.getString("host").getOrElse("dev")
-  private val http = new OkHttpClient()
-  private val appJson = MediaType.parse("application/json")
+  private val actions = new AddAssetActions(aws, hmac)
 
   override def addAsset(atomId: String, videoId: String): Long = {
-    val uri = s"https://$domain/api2/atoms/$atomId/assets"
-    val hmacHeaders = hmac.generateHmacHeaders(uri)
-
-    val videoUri = s"https://www.youtube.com/watch?v=$videoId"
-    val body = s"""{"uri": "$videoUri"}"""
-
-    if(aws.stage == "DEV") {
-      log.info(s"Add asset: POST $uri $body $hmacHeaders")
-      -1
-    } else {
-      val request = new Request.Builder()
-        .url(uri)
-        .headers(Headers.of(hmacHeaders.asJava))
-        .post(RequestBody.create(appJson, body))
-        .build()
-
-      val response = http.newCall(request).execute()
-      if (response.code() != 200) {
-        log.error(s"Unexpected response adding asset ${response.code()}")
-        log.error(s"uri=$uri body=$body responseBody=${response.body().string()}")
-        log.error(s"atomId=$atomId youTubeId=$videoId")
-
-        -1
-      } else {
-        val str = response.body().string()
-        val json = Json.parse(str)
-
-        parseAssetVersion(json)
-      }
-    }
-  }
-
-  private def parseAssetVersion(json: JsValue): Long = {
-    val assets = (json \ "assets").as[JsArray].value
-    val versions = assets.map { asset => (asset \ "version").as[Long] }
-
-    versions.sorted.last
+    actions.addAsset(atomId, videoId)
   }
 }
